Add tasks to install pip and docker-py module
The Ansible docker module (http://docs.ansible.com/ansible/docker_module.html)
has a host dependency on the docker-py python library, which by default is not
installed on Ubuntu. In order to be able to use the module to manage docker
containers, we need to first install the module.

The module is installed via pip in order to get the latest version, which may
not be provided in distribution apt repositories. See the following links for
more info:

 - https://github.com/ansible/ansible-modules-core/issues/1792
 - https://blog.confirm.ch/docker-ansible-on-rhel-centos-7/