Lay down the docker config file before installing docker engine
Due to the way Ansible triggers handlers, the 'notify docker reload' handler
was being executed after all roles were applied. In the case of the docker
registry, this meant that docker was not reloaded until after the registry
container had been started, which naturally caused the container to stop
again.

There were two solutions: either add a meta 'flush_handlers' statement to
force execution of the handlers, or simply get rid of the handlers altogether
since laying down the config file beforehand obviated the need to do a reload.